requirements.txt
streamlit
openai
python-dotenv
pypdf
faiss-cpu
sentence-transformers
numpy